Setting datum visibility prototype
When adding an item, you can prefix the datum names with underscores to make them private.
When editing an item, you can type "private", "public" or "internal" in the datums input boxes to set their visibility.

(Currently the first method doesn't work when editing items, and the second one doesn't work when adding them. This will be fixed in a future commit)