Add datum and template fields visibility (#7)
* Setting datum visibility prototype

When adding an item, you can prefix the datum names with underscores to make them private.
When editing an item, you can type "private", "public" or "internal" in the datums input boxes to set their visibility.

(Currently the first method doesn't work when editing items, and the second one doesn't work when adding them. This will be fixed in a future commit)

* Setting datum visibility

* Set visibility in template fields

* Added visibility selector to datum twig

* Added visibility to FieldFactory

* Database migration for adding visibility to template fields

* Changed "page"->"field" in fields' visibility descriptions

* templates: Fix styles and applying visibility for template fields

---------

Co-authored-by: Bartosz Leśniewski <blesniewski@outlook.com>